rename some files and outline AL ipynb
